+++ R/anglr.R
@@ -84,9 +84,10 @@ anglr <- function (x, z = NULL, ..., type = NULL, max_area = NULL) {
 #' @name anglr
 #' @export
 anglr.default <- function(x, z = NULL, ..., type = NULL, max_area = NULL) {
-  .Deprecated(new = "silicate::SC or silicate::TRI, and plot3d.SC", 
-              package = "anglr", 
-              old = "anglr(<line-topology>)")
+  .Defunct(new = "silicate::SC or silicate::TRI, and plot3d.SC", package = "anglr")
+  #.Deprecated(new = "silicate::SC or silicate::TRI, and plot3d.SC", 
+  #            package = "anglr", 
+  #            old = "anglr(<line-topology>)")
   ## we don't need methods for sf, Spatial etc because PATH covers those
   anglr(PATH(x), z = z, ..., type = type, max_area = max_area)
 }