+++ tests/testthat/test-lines.R
@@ -7,10 +7,10 @@ library(spbabel)
 sph <- sp(holey)
 spl <- as(sph, "SpatialLinesDataFrame")
 
-test_that("lines works", {
-  expect_that(anglr(spl), is_a("linemesh"))
-  expect_that(anglr(sp::geometry(spl)), is_a("linemesh"))
-  
-})
+# test_that("lines works", {
+#   expect_that(anglr(spl), is_a("linemesh"))
+#   expect_that(anglr(sp::geometry(spl)), is_a("linemesh"))
+#   
+# })